Get deals at a specific location using Groupon API
Find deals for the user's location using Groupon API and Google
locations and Google Maps API. Then a list view and a map view is
provided for all the deals on two separate tabs. The user can also
change the location using a drop down menu. A wishlist is provided to
store user specific deals. A login page is created to enter the App and
the user is authenticated based on his credentials stored in Parse.com